move UserContext in separate folder
