+++ src/components/LoginSection.js
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { UserContext } from "../util/UserContext";
+import { UserContext } from "../context/UserContext";
 import styled from "styled-components";
 
 export default function LoginSection() {
