+++ src/index.js
@@ -4,14 +4,17 @@ import GlobalStyle from "./GlobalStyle";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
+import { UserContextProvider } from "./context/UserContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <BrowserRouter basename="/">
-      <GlobalStyle />
-      <App />
-    </BrowserRouter>
+    <UserContextProvider>
+      <BrowserRouter basename="/">
+        <GlobalStyle />
+        <App />
+      </BrowserRouter>
+    </UserContextProvider>
   </React.StrictMode>
 );
